+++ packages/common/src/index.ts
@@ -36,6 +36,7 @@ export {
     deserializeInitError,
 } from './deserialization';
 export { getIdProof, IdStatementBuilder } from './idProofs';
+export * from './idProofTypes';
 export * from './signHelpers';
 export * from './accountHelpers';
 export * from './blockSummaryHelpers';
